requirements.txt
# LLM API
openai>=1.0.0

# 向量数据库
chromadb>=0.4.0

# 环境变量管理
python-dotenv>=1.0.0

# PDF 处理
pymupdf>=1.23.0

# Token 计数
tiktoken>=0.5.0
